Don't use responsive table on reports list
The responsive table actually does *worse* in narrow layouts because
the action column fucks off out of view and needs to be scrolled to
if the report reason is long. With a normal table you may not get
consistent row heights, but at least you don't have to horizontally
scroll to perform an action.

Also fixed a goof where I used a style attribute instead of making
the max-width for the action column a class.